pyproject.toml
[project]
name = "notirelevanteperu"
version = "0.1.0"
description = "Daily scraper + AI summarizer + bento dashboard for Peru's official gazette (El Peruano)"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "pydantic>=2.7",
    "google-genai>=0.8",
    "weasyprint>=62",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "tenacity>=8.4",
]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "ruff>=0.5",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

